src/requirements.txt
streamlit
openai==0.27.7
pytube
youtube-transcript-api
langchain==0.1.12
fastapi==0.79.0
uvicorn==0.18.2
python-dotenv==0.20.0
